mysql_api/requirements.txt
flask
werkzeug
flask-cors
mysql-connector-python